# public/docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Lumen: Loja de vendas",
        "version": "0.1"
    },
    "paths": {
...
        "/products/create": {
            "post": {
                "tags": [
                    "Produto"
                ],
                "summary": "Cadastrar novo produto",
                "description": "OBS: id do fabricante deve ser válido!",
                "operationId": "988c9dfe2fa2bd324bcf40e40d37d2ee",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_name",
                                    "product_description",
                                    "fabrication_date",
                                    "expiration_date",
                                    "buy_price",
                                    "sell_price",
                                    "manufacturer_id"
                                ],
                                "properties": {
                                    "product_name": {
                                        "type": "string",
                                        "example": "Farofa Zaeli"
                                    },
                                    "product_description": {
                                        "type": "string",
                                        "example": "A melhor do mercado!"
                                    },
                                    "fabrication_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-06-20"
                                    },
                                    "expiration_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-06-20"
                                    },
                                    "buy_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 19
                                    },
                                    "sell_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 28
                                    },
                                    "manufacturer_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Produto criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "product_name": {
                                            "type": "string",
                                            "example": "Farofa Zaeli"
                                        },
                                        "product_description": {
                                            "type": "string",
                                            "example": "A melhor do mercado!"
                                        },
                                        "fabrication_date": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2025-06-20"
                                        },
                                        "expiration_date": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2026-06-20"
                                        },
                                        "buy_price": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 19
                                        },
                                        "sell_price": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 28
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos"
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "tags": [
                    "Produto"
                ],
                "summary": "Buscar um produto",
                "operationId": "b2dc537014e0b7a417594278d020f99d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do produto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Aqui está o produto desejado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "product_name": {
                                            "type": "string",
                                            "example": "Farofa Zaeli"
                                        },
                                        "product_description": {
                                            "type": "string",
                                            "example": "A melhor do mercado!"
                                        },
                                        "fabrication_date": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2025-06-20"
                                        },
                                        "expiration_date": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2026-06-20"
                                        },
                                        "buy_price": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 19
                                        },
                                        "sell_price": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 28
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos"
                    }
                }
            }
        },
        "/products/update/{id}": {
            "put": {
                "tags": [
                    "Produto"
                ],
                "summary": "Alterar um cadastro",
                "operationId": "dd3708926693d9d189fbaaccbb852a1c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do produto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_name",
                                    "product_description",
                                    "sell_price",
                                    "buy_price"
                                ],
                                "properties": {
                                    "product_name": {
                                        "type": "string",
                                        "example": "Farofa Zaeli"
                                    },
                                    "product_description": {
                                        "type": "string",
                                        "example": "A melhor do mercado!"
                                    },
                                    "buy_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 19
                                    },
                                    "sell_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 28
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produto alterado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "product_name": {
                                            "type": "string",
                                            "example": "Farofa Pinduca"
                                        },
                                        "product_description": {
                                            "type": "string",
                                            "example": "A mais saborosa."
                                        },
                                        "buy_price": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 32
                                        },
                                        "sell_price": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 48
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos"
                    }
                }
            }
        },
        "/products/delete/{id}": {
            "delete": {
                "tags": [
                    "Produto"
                ],
                "summary": "Deletar um produto pelo ID",
                "operationId": "a79c11341007437b9eba1cec11144066",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do produto a ser deletado.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Produto deletado com sucesso"
                    },
                    "404": {
                        "description": "Produto não encontrado"
                    }
                }
            }
        },
...
    },
...
}